server/requirements.txt
Flask
flask-cors
mongoengine
python-dotenv
flask_jwt_extended
flask_bcrypt 